fix(US-002): advance offset only past complete lines; drive tests by observation
Two adversarial-review findings from US-002 are addressed:

- consumeRange now advances the byte offset only to the position past the
  last complete line (last \\n). Bytes after that form a partial line that
  the next poll re-reads, so a writer that flushes mid-line and completes
  on a subsequent poll does not silently lose the entry. Updated the
  followLogs doc-comment to call out the new partial-line-continuation
  invariant.

- All seven US-002 tests drop the fixed-duration real-timer races
  (await new Promise(r => setTimeout(r, N))) in favour of waitForCondition()
  polling a real observable side effect (emitted[] or readRangeCalls[]) and
  withTimeout() bounding the followLogs promise. Adds a partial-line
  regression test, and updates the logs-formatter US-001 fixture to use a
  realistic JSONL trailing newline so the AC#3 cancellation test stays in
  scope after the new last-\\n boundary.

Split the seven byte-offset tests into test/unit/commands/logs-follow-byte.test.ts
to keep logs.test.ts under the 800-line test-file limit per
.claude/rules/test-architecture.md.
