+++ test/unit/commands/logs-formatter.test.ts
@@ -50,7 +50,8 @@ const SAMPLE_ENTRIES = [
 
 function writeFixture(dir: string, name: string, entries: typeof SAMPLE_ENTRIES): string {
   const path = join(dir, name);
-  writeFileSync(path, entries.map((e) => JSON.stringify(e)).join("\n"));
+  // Realistic JSONL: each line is newline-terminated.
+  writeFileSync(path, entries.map((e) => JSON.stringify(e)).join("\n") + "\n");
   return path;
 }
