+++ test/unit/commands/logs.test.ts
@@ -14,16 +14,9 @@
  */
 
 import { afterEach, beforeEach, describe, expect, test } from "bun:test";
-import { existsSync, mkdirSync, rmSync, writeFileSync, appendFileSync } from "node:fs";
+import { existsSync, mkdirSync, rmSync, writeFileSync } from "node:fs";
 import { join } from "node:path";
-import { cleanupTempDir, makeTempDir } from "@test/helpers";
-import {
-  type FollowLogsDeps,
-  type LogsOptions,
-  _deps,
-  followLogs,
-  logsCommand,
-} from "../../../src/commands/logs";
+import { type LogsOptions, _deps, logsCommand } from "../../../src/commands/logs";
 
 const TEST_WORKSPACE = join(import.meta.dir, "..", "..", "tmp", "logs-test");
 
@@ -228,265 +221,6 @@ describe("logsCommand", () => {
       const result = logsCommand({ dir: projectDir, follow: true, signal: controller.signal });
       controller.abort();
       await expect(result).resolves.toBeUndefined();
-    });
-  });
-
-  describe("--follow mode byte-offset tailing (US-002)", () => {
-    let followDir: string;
-
-    beforeEach(() => {
-      followDir = makeTempDir("nax-logs-follow-byte-");
-    });
-
-    afterEach(() => {
-      cleanupTempDir(followDir);
-    });
-
-    function makeEntry(message: string, storyId = "US-001"): Record<string, unknown> {
-      return {
-        timestamp: "2026-02-27T10:00:00.000Z",
-        level: "info",
-        stage: "test",
-        storyId,
-        message,
-        data: {},
-      };
-    }
-
-    function encodeLine(entry: Record<string, unknown>): Buffer {
-      return Buffer.from(`${JSON.stringify(entry)}\n`, "utf8");
-    }
-
-    async function writeInitial(filePath: string, entry: Record<string, unknown>): Promise<void> {
-      await Bun.write(filePath, encodeLine(entry));
-    }
-
-    async function appendEntry(filePath: string, entry: Record<string, unknown>): Promise<void> {
-      appendFileSync(filePath, encodeLine(entry));
-    }
-
-    async function rewriteShorter(filePath: string, entry: Record<string, unknown>): Promise<void> {
-      await Bun.write(filePath, encodeLine(entry));
-    }
-
-    test("emits appended entry intact when preceding message contains three ✓ characters (AC #1)", async () => {
-      const filePath = join(followDir, "run.jsonl");
-      await writeInitial(filePath, makeEntry("✓ ✓ ✓"));
-      const appended = makeEntry("appended entry");
-
-      const emitted: string[] = [];
-      let appendedOnce = false;
-      const deps: FollowLogsDeps = {
-        emit: (line) => emitted.push(line),
-        sleep: async () => {
-          if (!appendedOnce) {
-            await appendEntry(filePath, appended);
-            appendedOnce = true;
-          }
-        },
-        readRange: async (path, start) =>
-          Buffer.from(await Bun.file(path).slice(start).arrayBuffer()).toString("utf8"),
-      };
-      const controller = new AbortController();
-
-      const follow = followLogs(filePath, {}, { signal: controller.signal, _deps: deps });
-      // Let the loop run at least one poll cycle, then cancel.
-      await new Promise((r) => setTimeout(r, 10));
-      controller.abort();
-      await follow;
-
-      // Pre-existing entry emitted once on initial read; appended entry emitted
-      // once on poll. The byte-aligned read must NOT silently drop the append.
-      expect(emitted.length).toBe(2);
-      expect(emitted[0]).toContain("✓ ✓ ✓");
-      expect(emitted[1]).toContain("appended entry");
-    });
-
-    test("emits three successive ✓-containing entries across three polls in append order (AC #2)", async () => {
-      const filePath = join(followDir, "run.jsonl");
-      await writeInitial(filePath, makeEntry("seed ✓"));
-
-      const appended = [
-        makeEntry("first ✓"),
-        makeEntry("second ✓"),
-        makeEntry("third ✓"),
-      ];
-      let sleepCalls = 0;
-
-      const emitted: string[] = [];
-      const deps: FollowLogsDeps = {
-        emit: (line) => emitted.push(line),
-        sleep: async () => {
-          if (sleepCalls < appended.length) {
-            await appendEntry(filePath, appended[sleepCalls]);
-          }
-          sleepCalls++;
-        },
-        readRange: async (path, start) =>
-          Buffer.from(await Bun.file(path).slice(start).arrayBuffer()).toString("utf8"),
-      };
-      const controller = new AbortController();
-
-      const follow = followLogs(filePath, {}, { signal: controller.signal, _deps: deps });
-      // Let the loop run three polls, then cancel.
-      await new Promise((r) => setTimeout(r, 10));
-      controller.abort();
-      await follow;
-
-      expect(sleepCalls).toBeGreaterThanOrEqual(3);
-      // Each appended ✓ entry must be emitted exactly once and in append order.
-      const firstIdx = emitted.findIndex((l) => l.includes("first ✓"));
-      const secondIdx = emitted.findIndex((l) => l.includes("second ✓"));
-      const thirdIdx = emitted.findIndex((l) => l.includes("third ✓"));
-      expect(firstIdx).toBeGreaterThanOrEqual(0);
-      expect(secondIdx).toBeGreaterThan(firstIdx);
-      expect(thirdIdx).toBeGreaterThan(secondIdx);
-      // Each appended entry appears exactly once (no double-emit on re-poll).
-      expect(emitted.filter((l) => l.includes("first ✓")).length).toBe(1);
-      expect(emitted.filter((l) => l.includes("second ✓")).length).toBe(1);
-      expect(emitted.filter((l) => l.includes("third ✓")).length).toBe(1);
-    });
-
-    test("emits every appended entry exactly once across pre-existing + two appends (AC #3)", async () => {
-      const filePath = join(followDir, "run.jsonl");
-      await writeInitial(filePath, makeEntry("pre-existing ✓"));
-
-      const appendedA = makeEntry("append A ✓");
-      const appendedB = makeEntry("append B ✓");
-      let sleepCalls = 0;
-
-      const emitted: string[] = [];
-      const deps: FollowLogsDeps = {
-        emit: (line) => emitted.push(line),
-        sleep: async () => {
-          sleepCalls++;
-          if (sleepCalls === 1) {
-            await appendEntry(filePath, appendedA);
-          } else if (sleepCalls === 2) {
-            await appendEntry(filePath, appendedB);
-          }
-        },
-        readRange: async (path, start) =>
-          Buffer.from(await Bun.file(path).slice(start).arrayBuffer()).toString("utf8"),
-      };
-      const controller = new AbortController();
-
-      const follow = followLogs(filePath, {}, { signal: controller.signal, _deps: deps });
-      await new Promise((r) => setTimeout(r, 10));
-      controller.abort();
-      await follow;
-
-      const aCount = emitted.filter((l) => l.includes("append A ✓")).length;
-      const bCount = emitted.filter((l) => l.includes("append B ✓")).length;
-      expect(aCount).toBe(1);
-      expect(bCount).toBe(1);
-    });
-
-    test("skips an invalid JSON line between two ✓ entries and emits the later valid entry (AC #4)", async () => {
-      const filePath = join(followDir, "run.jsonl");
-      await writeInitial(filePath, makeEntry("seed ✓"));
-
-      const validA = makeEntry("valid A ✓");
-      const validB = makeEntry("valid B ✓");
-      let sleepCalls = 0;
-
-      const emitted: string[] = [];
-      const deps: FollowLogsDeps = {
-        emit: (line) => emitted.push(line),
-        sleep: async () => {
-          sleepCalls++;
-          if (sleepCalls === 1) {
-            appendFileSync(filePath, encodeLine(validA));
-            appendFileSync(filePath, Buffer.from("this is not valid json\n", "utf8"));
-            appendFileSync(filePath, encodeLine(validB));
-          }
-        },
-        readRange: async (path, start) =>
-          Buffer.from(await Bun.file(path).slice(start).arrayBuffer()).toString("utf8"),
-      };
-      const controller = new AbortController();
-
-      const follow = followLogs(filePath, {}, { signal: controller.signal, _deps: deps });
-      await new Promise((r) => setTimeout(r, 10));
-      controller.abort();
-      await follow;
-
-      const aCount = emitted.filter((l) => l.includes("valid A ✓")).length;
-      const bCount = emitted.filter((l) => l.includes("valid B ✓")).length;
-      expect(aCount).toBe(1);
-      expect(bCount).toBe(1);
-    });
-
-    test("resynchronises when the file is rewritten shorter than the consumed offset, then emits the next append (AC #5)", async () => {
-      const filePath = join(followDir, "run.jsonl");
-      await writeInitial(filePath, makeEntry("seed ✓ ✓ ✓ ✓ ✓"));
-
-      const truncatedEntry = makeEntry("after-truncation ✓");
-      const appendedEntry = makeEntry("post-resync ✓");
-      let sleepCalls = 0;
-
-      const emitted: string[] = [];
-      const deps: FollowLogsDeps = {
-        emit: (line) => emitted.push(line),
-        sleep: async () => {
-          sleepCalls++;
-          if (sleepCalls === 1) {
-            await rewriteShorter(filePath, truncatedEntry);
-          } else if (sleepCalls === 2) {
-            await appendEntry(filePath, appendedEntry);
-          }
-        },
-        readRange: async (path, start) =>
-          Buffer.from(await Bun.file(path).slice(start).arrayBuffer()).toString("utf8"),
-      };
-      const controller = new AbortController();
-
-      const follow = followLogs(filePath, {}, { signal: controller.signal, _deps: deps });
-      await new Promise((r) => setTimeout(r, 10));
-      controller.abort();
-      await follow;
-
-      // The appended entry after truncation must be emitted.
-      const appendedCount = emitted.filter((l) => l.includes("post-resync ✓")).length;
-      expect(appendedCount).toBe(1);
-    });
-
-    test("readRange receives start offset equal to bytes already consumed, not 0, on append polls (AC #6)", async () => {
-      const filePath = join(followDir, "run.jsonl");
-      const initial = makeEntry("seed ✓");
-      await writeInitial(filePath, initial);
-      const initialSize = encodeLine(initial).length;
-
-      let sleepCalls = 0;
-      const readRangeCalls: number[] = [];
-      const deps: FollowLogsDeps = {
-        emit: () => {},
-        sleep: async () => {
-          sleepCalls++;
-          if (sleepCalls === 1) {
-            await appendEntry(filePath, makeEntry("append ✓"));
-          }
-        },
-        readRange: async (_path, start) => {
-          readRangeCalls.push(start);
-          return Buffer.from(await Bun.file(filePath).slice(start).arrayBuffer()).toString(
-            "utf8",
-          );
-        },
-      };
-      const controller = new AbortController();
-
-      const follow = followLogs(filePath, {}, { signal: controller.signal, _deps: deps });
-      await new Promise((r) => setTimeout(r, 5));
-      controller.abort();
-      await follow;
-
-      // The first call (during the initial read) is start=0.
-      // The second call (during the append poll) must equal the byte length
-      // already consumed — which is the size of the initial file.
-      expect(readRangeCalls.length).toBeGreaterThanOrEqual(2);
-      expect(readRangeCalls[0]).toBe(0);
-      expect(readRangeCalls[1]).toBe(initialSize);
     });
   });
